Add alt,width,height attributes to contact icon
Specifying the dimensions of the image as HTML attributes helps to
prevent layout shifts during the page loading phase. 300(px) was chosen
because that's the size of the viewBox specified in the SVG image's
markup. It has no relation to the actual apparent size of the icon to
the user.

Specifying a text alternative allows users of screen readers to
understand the meaning of the link.