Add #top for the "Top of page" link
The previous implementation of the "Top of page" link didn't move the
keyboard focus, so pressing Tab after navigating to the # anchor would
scroll the page back to the bottom.

By providing a real target (the top of `<body>`), the scroll position
and the keyboard focus position are both updated.